QPPA-1585: prevent sharedLogger.configure() being called twice. (#22)
- Add a check to configure() that warns if it is called more than once
and ignores the 2nd and subsequent calls.

When implementing QPPA-1585 we noticed that configure() was not working
as expected. We tracked this down to a library that was also calling
configure, after the main code called it. The settings for the 2nd call
were overwriting the first. In particular, redactKeys was lost in our
case.

This change prevents this from happening.