Add "null object" loggers for "none" logging (#32)
Under the old implementation, setting the logLevel or accessLog.format to 'none' would cause the calling program to crash if it called any log methods, since it would not define the logger object. Under this approach, it defines objects with empty function bodies instead that do nothing and thus suppress all logging output.